Fix bug with secure boot install

+++ arch/install.sh
@@ -173,19 +173,24 @@ msg "${PURPLE}\n=== Building needed AUR packages with temporary user, then insta
 arch-chroot /mnt /bin/bash <<EOF
 useradd -m build && passwd -d build && groupadd -rf wheel && gpasswd -a build wheel
 echo "%wheel ALL=(ALL) ALL" > /etc/sudoers.d/override
+
 chgrp build /home/build && chmod g+ws /home/build
 setfacl -m u::rwx,g::rwx /home/build && \
 setfacl -d --set u::rwx,g::rwx,o::- /home/build
 cd /home/build
-git clone https://aur.archlinux.org/arch-secure-boot.git && \
-cd arch-secure-boot && \
-sudo -u build makepkg -s --noconfirm --skippgpcheck && \
-pacman -U --noconfirm arch-secure-boot*pkg* && \ cd ..
+
+[[ -d /home/build ]] || exit 1
+
 git clone https://aur.archlinux.org/mkinitcpio-encrypt-detached-header.git && \
 cd mkinitcpio-encrypt-detached-header && \
 sudo -u build makepkg -src --noconfirm --skippgpcheck && \
-pacman -U --noconfirm mkinitcpio-encrypt-detached-header*pkg* && \
-cd .. && \
+pacman -U --noconfirm mkinitcpio-encrypt-detached-header*pkg* && cd ..
+
+git clone https://aur.archlinux.org/arch-secure-boot.git && \
+cd arch-secure-boot && \
+sudo -u build makepkg -s --noconfirm --skippgpcheck && \
+pacman -U --noconfirm arch-secure-boot*pkg* && cd ..
+
 userdel -r build && groupdel wheel && rm -rf /home/build
 EOF
